refactor requirements and test CheckCommand func
