+++ internal/commands/deploy/deploy.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hazelops/ize/internal/manager/alias"
 	"github.com/hazelops/ize/internal/manager/ecs"
 	"github.com/hazelops/ize/internal/manager/serverless"
+	"github.com/hazelops/ize/internal/requirements"
 	"github.com/hazelops/ize/pkg/templates"
 	"github.com/hazelops/ize/pkg/terminal"
 	"github.com/spf13/cobra"
@@ -96,12 +97,12 @@ func (o *Options) Complete(cmd *cobra.Command) error {
 		return fmt.Errorf("can't deploy your stack: %w", err)
 	}
 
-	if err = config.CheckRequirements(config.WithIzeStructure(), config.WithConfigFile()); err != nil {
+	if err = requirements.CheckRequirements(requirements.WithIzeStructure(), requirements.WithConfigFile()); err != nil {
 		return err
 	}
 
-	if o.Config.Serverless != nil {
-		if err = config.CheckRequirements(config.WithNVM()); err != nil {
+	if len(o.Config.Serverless) != 0 {
+		if err = requirements.CheckRequirements(requirements.WithNVM()); err != nil {
 			return err
 		}
 	}
