+++ internal/commands/down/down_infra.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/hazelops/ize/internal/config"
+	"github.com/hazelops/ize/internal/requirements"
 	"github.com/hazelops/ize/pkg/terminal"
 	"github.com/spf13/cobra"
 )
@@ -64,7 +65,7 @@ func (o *DownInfraOptions) Complete() error {
 		return fmt.Errorf("can't load options for a command: %w", err)
 	}
 
-	if err := config.CheckRequirements(config.WithIzeStructure(), config.WithConfigFile()); err != nil {
+	if err := requirements.CheckRequirements(requirements.WithIzeStructure(), requirements.WithConfigFile()); err != nil {
 		return err
 	}
