+++ internal/config/config.go
@@ -1,17 +1,16 @@
 package config
 
 import (
-	"errors"
 	"fmt"
 	"github.com/hazelops/ize/internal/schema"
 	"github.com/mitchellh/mapstructure"
+	"github.com/spf13/cobra"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"reflect"
 	"strings"
 
-	"github.com/Masterminds/semver"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/hazelops/ize/internal/aws/utils"
@@ -33,91 +32,6 @@ type Config struct {
 	IsGlobal        bool
 	IsDockerRuntime bool
 	IsPlainText     bool
-}
-
-type requiments struct {
-	configFile bool
-	smplugin   bool
-	structure  bool
-	nvm        bool
-}
-
-type Option func(*requiments)
-
-func WithIzeStructure() Option {
-	return func(r *requiments) {
-		r.structure = true
-	}
-}
-
-func WithConfigFile() Option {
-	return func(r *requiments) {
-		r.configFile = true
-	}
-}
-
-func WithSSMPlugin() Option {
-	return func(r *requiments) {
-		r.smplugin = true
-	}
-}
-
-func WithNVM() Option {
-	return func(r *requiments) {
-		r.nvm = true
-	}
-}
-
-func CheckRequirements(options ...Option) error {
-	r := requiments{}
-	for _, opt := range options {
-		opt(&r)
-	}
-
-	if r.nvm {
-		err := checkNVM()
-		if err != nil {
-			return err
-		}
-	}
-
-	if r.structure {
-		if !isStructured() {
-			pterm.Warning.Println("is not an ize-structured directory. Please run ize init or cd into an ize-structured directory.")
-		}
-	}
-
-	if r.smplugin {
-		if err := checkSessionManagerPlugin(); err != nil {
-			return err
-		}
-	}
-
-	switch viper.GetString("prefer_runtime") {
-	case "native":
-		logrus.Debug("use native runtime")
-	case "docker":
-		if err := checkDocker(); err != nil {
-			return err
-		}
-		logrus.Debug("use docker runtime")
-	default:
-		return fmt.Errorf("unknown runtime type: %s", viper.GetString("prefer_runtime"))
-	}
-
-	if len(viper.ConfigFileUsed()) == 0 && r.configFile {
-		return fmt.Errorf("this command required config file")
-	}
-
-	return nil
-}
-
-func checkNVM() error {
-	if len(os.Getenv("NVM_DIR")) == 0 {
-		return errors.New("nvm is not installed (visit https://github.com/nvm-sh/nvm)")
-	}
-
-	return nil
 }
 
 func GetConfig() (*Project, error) {
@@ -299,112 +213,6 @@ func setDefaultInfraDir(cwd string) {
 	}
 }
 
-func checkDocker() error {
-	_, err := CheckCommand("docker", []string{"info"})
-	if err != nil {
-		return errors.New("docker is not running or is not installed (visit https://www.docker.com/get-started)")
-	}
-
-	return nil
-}
-
-func isStructured() bool {
-	var isStructured = false
-
-	cwd, err := os.Getwd()
-	if err != nil {
-		logrus.Fatalln("can't initialize config: %w", err)
-	}
-
-	_, err = os.Stat(filepath.Join(cwd, ".ize"))
-	if !os.IsNotExist(err) {
-		isStructured = true
-	}
-
-	_, err = os.Stat(filepath.Join(cwd, ".infra"))
-	if !os.IsNotExist(err) {
-		isStructured = true
-	}
-
-	return isStructured
-}
-
-func checkSessionManagerPlugin() error {
-	_, err := CheckCommand("session-manager-plugin", []string{})
-	if err != nil {
-		pterm.Warning.Println("SSM Agent plugin is not installed. Trying to install SSM Agent plugin")
-
-		var pyVersion string
-
-		pyVersion, err = CheckCommand("python3", []string{"--version"})
-		if err != nil {
-			pyVersion, err = CheckCommand("python", []string{"--version"})
-			if err != nil {
-				return errors.New("python is not installed")
-			}
-
-			c, err := semver.NewConstraint("<= 2.6.5")
-			if err != nil {
-				return err
-			}
-
-			v, err := semver.NewVersion(strings.TrimSpace(strings.Split(pyVersion, " ")[1]))
-			if err != nil {
-				return err
-			}
-
-			if c.Check(v) {
-				return fmt.Errorf("python version %s below required %s", v.String(), "2.6.5")
-			}
-			return errors.New("python is not installed")
-		}
-
-		c, err := semver.NewConstraint("<= 3.3.0")
-		if err != nil {
-			return err
-		}
-
-		v, err := semver.NewVersion(strings.TrimSpace(strings.Split(pyVersion, " ")[1]))
-		if err != nil {
-			return err
-		}
-
-		if c.Check(v) {
-			return fmt.Errorf("python version %s below required %s", v.String(), "3.3.0")
-		}
-
-		pterm.DefaultSection.Println("Installing SSM Agent plugin")
-
-		err = downloadSSMAgentPlugin()
-		if err != nil {
-			return fmt.Errorf("download SSM Agent plugin error: %v (visit https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html)", err)
-		}
-
-		pterm.Success.Println("Downloading SSM Agent plugin")
-
-		err = installSSMAgent()
-		if err != nil {
-			return fmt.Errorf("install SSM Agent plugin error: %v (visit https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html)", err)
-		}
-
-		pterm.Success.Println("Installing SSM Agent plugin")
-
-		err = cleanupSSMAgent()
-		if err != nil {
-			return fmt.Errorf("cleanup SSM Agent plugin error: %v (visit https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html)", err)
-		}
-
-		pterm.Success.Println("Cleanup Session Manager plugin installation package")
-
-		_, err = CheckCommand("session-manager-plugin", []string{})
-		if err != nil {
-			return fmt.Errorf("check SSM Agent plugin error: %v (visit https://docs.aws.amazon.com/systems-manager/latest/userguide/session-manager-working-with-install-plugin.html)", err)
-		}
-	}
-
-	return nil
-}
-
 func readGlobalConfigFile() (*Config, error) {
 	env := viper.GetString("env")
 	namespace := viper.GetString("namespace")
@@ -560,4 +368,27 @@ func structToMap(item interface{}) map[string]interface{} {
 	}
 
 	return res
+}
+
+//GetApps returns a list of application names in the directory for shell completions
+func GetApps(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	var apps []string
+
+	dir, _ := os.ReadDir("./apps")
+
+	if dir != nil {
+		for _, entry := range dir {
+			apps = append(apps, entry.Name())
+		}
+	}
+
+	dir, _ = os.ReadDir("./projects")
+
+	if dir != nil {
+		for _, entry := range dir {
+			apps = append(apps, entry.Name())
+		}
+	}
+
+	return apps, cobra.ShellCompDirectiveNoFileComp
 }
